refactor -> use datetime.fromisoformat
After trying several different DateTime parsers, I found out the issue was with the query I was using for testing. Sending a timestamp with a '+' inside of it in a get request without encoding results in an incorrect DateTime string. Changing the + to its encoded form of %2B fixed this issue.
Also removed the convert_timezone function as .astimezone is easy enough to read on its own.