refactor -> submit_date changes
Being that the idempotency key is being used as essentially a log number, changed from update_or_create to simply create. This saves all of the submitted reviews (aside from those with an idential idempotency key) into the DB.
Also, with the addition of submit_date to the ReviewResult object, the filtered cards .order_by function has now been changed to use this field.